Divided the classes' tests into different files
